+++ application/libraries/Miscelleneous.php
@@ -3297,7 +3297,10 @@ function generate_image($base64, $image_name,$directory){
         header('Content-Length: ' . filesize($csv));
         readfile($csv);
         exec("rm -rf " . escapeshellarg($csv));
-        unlink($csv);
+        if(file_exists($csv))
+        {
+            unlink($csv);
+        }
     }
     /*
      * This Function used to load navigation header from cache
@@ -3390,7 +3393,10 @@ function generate_image($base64, $image_name,$directory){
                    // $login_email['password'] = "***********";
                     $login_emailBody247 = vsprintf($login_template[0], $login_email);
                     //Send Login Details to partner
-                    $this->My_CI->notify->sendEmail($login_template[2], $data['email'], "", "",$login_subject, $login_emailBody, "",'partner_login_details');
+                    if(!empty($data['email']))
+                    {
+                        $this->My_CI->notify->sendEmail($login_template[2], $data['email'], "", "",$login_subject, $login_emailBody, "",'partner_login_details');
+                    }
                     //Send Login Details to 247around 
                     $to = $this->My_CI->session->userdata('official_email');
                     $this->My_CI->notify->sendEmail($login_template[2], $to, $cc, $bcc,$login_subject, $login_emailBody247, "",'partner_login_details');
@@ -4481,12 +4487,28 @@ function generate_image($base64, $image_name,$directory){
      * @desc This function is used to process spare transfer
      */
     function spareTransfer($bookings_spare, $agentid, $agent_name, $login_partner_id, $login_service_center_id) {
+        
+       
+        if ($this->My_CI->session->userdata('emp_name') && $this->My_CI->session->userdata('userType') != 'partner') {
+            $agent_id = $this->My_CI->session->userdata('id');
+            $track_entity_type = _247AROUND_EMPLOYEE_STRING;
+            $track_partner_id = _247AROUND;
+        } else if ($this->My_CI->session->userdata('userType') == 'service_center') {
+            $agentid = $this->My_CI->session->userdata('service_center_agent_id');
+            $track_partner_id = $this->My_CI->session->userdata('service_center_id');
+            $track_entity_type = _247AROUND_SF_STRING;
+        } else {
+            $agentid = '';
+            $track_partner_id = '';
+            $track_entity_type = '';
+        }
+
         $tcount = 0;
         $booking_error_array = array();
         $add_row = array();
         
         foreach ($bookings_spare as $booking) {
-            $spareid = $booking['id'];
+            $spare_id = $spareid = $booking['id'];
             $partner_id = $booking['partner_id'];
             $state = $booking['state'];
             $delivered_sp=array();
@@ -4552,10 +4574,14 @@ function generate_image($base64, $image_name,$directory){
                            
                             $this->My_CI->inventory_model->update_pending_inventory_stock_request(_247AROUND_SF_STRING, $data['entity_id'], $data['inventory_id'], $booking['quantity']);
                         }
+                        /* Insert Spare Tracking Details */
+                        if (!empty($spare_id)) {
+                            $tracking_details = array('spare_id' => $spare_id, 'action' => $new_state, 'remarks' => trim($remarks), 'agent_id' => $agent_id, 'entity_id' => $track_partner_id, 'entity_type' => $track_entity_type);
+                            $this->My_CI->service_centers_model->insert_spare_tracking_details($tracking_details);
+                        }
 
 
-
-                         $this->My_CI->notify->insert_state_change($booking['booking_id'], $new_state, $old_state, $remarks, $agentid,$agent_name, $actor, $next_action, $login_partner_id, $login_service_center_id);
+                        $this->My_CI->notify->insert_state_change($booking['booking_id'], $new_state, $old_state, $remarks, $agentid,$agent_name, $actor, $next_action, $login_partner_id, $login_service_center_id, $spare_id);
 
 
                         if ($data['is_micro_wh']==1) {
@@ -4578,7 +4604,12 @@ function generate_image($base64, $image_name,$directory){
                         $remarks = _247AROUND_TRANSFERED_TO_PARTNER;
 				        
                         $actor="Partner";
-                        $this->My_CI->notify->insert_state_change($booking['booking_id'], $new_state, $old_state, $remarks, $agentid,$agent_name, $actor, $next_action, $login_partner_id, $login_service_center_id);
+                        /* Insert Spare Tracking Details */
+                        if (!empty($spare_id)) {
+                            $tracking_details = array('spare_id' => $spare_id, 'action' => $new_state, 'remarks' => trim($remarks), 'agent_id' => $agentid, 'entity_id' => $track_partner_id, 'entity_type' => $track_entity_type);
+                            $this->My_CI->service_centers_model->insert_spare_tracking_details($tracking_details);
+                        }
+                        $this->My_CI->notify->insert_state_change($booking['booking_id'], $new_state, $old_state, $remarks, $agentid,$agent_name, $actor, $next_action, $login_partner_id, $login_service_center_id, $spare_id);
                         $this->My_CI->inventory_model->update_pending_inventory_stock_request(_247AROUND_SF_STRING, $partner_id, $requested_inventory, -$booking['quantity']);
                     }
                     $tcount++;
@@ -5097,4 +5128,16 @@ function generate_image($base64, $image_name,$directory){
         }
         return $arr_emp;
     }
+
+/*  Getting URL called  Abhishek Awasthi*/
+
+    function get_uri_called(){
+        $base = base_url();
+        $controller = $this->My_CI->router->fetch_class();
+        $method = $this->My_CI->router->fetch_method();
+        return $base."/".$controller."/".$method;
+    }
+
+
+
 }
